Move notification functions to common
The package init should contain only functions needed for the package
initialization. All other functions need to be in different files such
as common or utils or any more specific file.
