+++ test/integration/ggrc_workflows/notifications/test_cycle_start_failed_notifications.py
@@ -8,7 +8,7 @@ from freezegun import freeze_time
 from datetime import datetime
 from mock import patch
 
-from ggrc import notifications
+from ggrc.notifications import common
 from ggrc.models import Notification
 from integration.ggrc_workflows.generator import WorkflowsGenerator
 from integration.ggrc.api_helper import Api
@@ -55,24 +55,24 @@ class TestCycleStartFailed(TestCase):
       self.assert200(response)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(wf_owner, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(wf_owner, notif_data)
       self.assertIn("cycle_starts_in", notif_data[wf_owner])
 
     with freeze_time("2015-03-05 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(wf_owner, notif_data)
       self.assertNotIn("cycle_started", notif_data[wf_owner])
       self.assertIn(wf_owner, notif_data)
       self.assertIn("cycle_start_failed", notif_data[wf_owner])
 
-      notifications.send_todays_digest_notifications()
+      common.send_todays_digest_notifications()
 
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(wf_owner, notif_data)
 
   # TODO: investigate why next_cycle_start date remains the same after
@@ -91,31 +91,31 @@ class TestCycleStartFailed(TestCase):
   #     self.assert200(response)
 
   #   with freeze_time("2015-01-01 13:39:20"):
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
   #   with freeze_time("2015-01-29 13:39:20"):
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_starts_in", notif_data[wf_owner])
 
   #   with freeze_time("2015-02-05 13:39:20"):
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_start_failed", notif_data[wf_owner])
 
   #     start_recurring_cycles()
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_start_failed", notif_data[wf_owner])
 
-  #     notifications.send_todays_digest_notifications()
+  #     common.send_todays_digest_notifications()
 
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
   # @patch("ggrc.notifications.common.send_email")
@@ -133,22 +133,22 @@ class TestCycleStartFailed(TestCase):
 
   #     self.assert200(response)
 
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_start_failed", notif_data[wf_owner])
 
   #     start_recurring_cycles()
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_start_failed", notif_data[wf_owner])
 
-  #     notifications.send_todays_digest_notifications()
+  #     common.send_todays_digest_notifications()
 
-  #     _, notif_data = notifications.get_todays_notifications()
+  #     _, notif_data = common.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
   def create_test_cases(self):
