+++ test/integration/ggrc_workflows/notifications/test_enable_disable_notifications.py
@@ -8,8 +8,7 @@ from freezegun import freeze_time
 from datetime import datetime
 from mock import patch
 
-from ggrc import notifications
-from ggrc.models import Notification, Person
+from ggrc.notifications import common
 from integration.ggrc_workflows.generator import WorkflowsGenerator
 from integration.ggrc.api_helper import Api
 from integration.ggrc.generator import ObjectGenerator
@@ -54,11 +53,11 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
   @patch("ggrc.notifications.common.send_email")
@@ -76,11 +75,11 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
   @patch("ggrc.notifications.common.send_email")
@@ -93,14 +92,14 @@ class TestEnableAndDisableNotifications(TestCase):
 
     with freeze_time("2015-01-29 13:39:20"):
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
       self.object_generator.generate_notification_setting(
           self.user.id, "Email_Digest", True)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
   @patch("ggrc.notifications.common.send_email")
@@ -115,14 +114,14 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
       self.object_generator.generate_notification_setting(
           self.user.id, "Email_Digest", True)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
   @patch("ggrc.notifications.common.send_email")
@@ -140,7 +139,7 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[user.email])
       self.assertIn(wf_forced.id, notif_data[user.email]["cycle_starts_in"])
@@ -150,7 +149,7 @@ class TestEnableAndDisableNotifications(TestCase):
           self.user.id, "Email_Digest", False)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[user.email])
       self.assertIn(wf_forced.id, notif_data[user.email]["cycle_starts_in"])
