+++ test/integration/ggrc_workflows/notifications/test_monthly_wf.py
@@ -8,7 +8,7 @@ from freezegun import freeze_time
 from datetime import datetime
 from mock import patch
 
-from ggrc import notifications
+from ggrc.notifications import common
 from ggrc.models import Notification, Person
 from ggrc_workflows import start_recurring_cycles
 from integration.ggrc_workflows.generator import WorkflowsGenerator
@@ -54,18 +54,18 @@ class TestMonthlyWorkflowNotification(TestCase):
       person_1 = Person.query.get(self.person_1.id)
 
     with freeze_time("2015-04-02"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertIn(person_1.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[person_1.email])
 
     with freeze_time("2015-04-02"):
       self.api.tc.get("nightly_cron_endpoint")
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(person_1.email, notif_data)
 
     with freeze_time("2015-04-02"):
       start_recurring_cycles()
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(person_1.email, notif_data)
 
     # cycle starts on monday - 6th, and not on 5th
@@ -73,12 +73,12 @@ class TestMonthlyWorkflowNotification(TestCase):
       start_recurring_cycles()
 
     with freeze_time("2015-04-15"):  # one day befor due date
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       person_1 = Person.query.get(self.person_1.id)
       self.assertIn(person_1.email, notif_data)
 
     with freeze_time("2015-04-25"):  # due date
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       person_1 = Person.query.get(self.person_1.id)
       self.assertIn(person_1.email, notif_data)
 
@@ -92,16 +92,16 @@ class TestMonthlyWorkflowNotification(TestCase):
       person_1 = Person.query.get(self.person_1.id)
 
     with freeze_time("2015-04-03"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
 
     with freeze_time("2015-04-03"):
       _, cycle = self.wf_generator.generate_cycle(wf)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       person_1 = Person.query.get(self.person_1.id)
       self.assertIn("cycle_started", notif_data[person_1.email])
 
     with freeze_time("2015-05-03"):  # two days befor due date
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       person_1 = Person.query.get(self.person_1.id)
       self.assertIn(person_1.email, notif_data)
