+++ test/integration/ggrc_workflows/notifications/test_one_time_wf_decline_accept_tasks.py
@@ -183,7 +183,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.wf_generator.activate_workflow(workflow)
 
     with freeze_time("2015-05-02"):
-      notifications.send_todays_digest_notifications()
+      common.send_todays_digest_notifications()
 
       cycle = Cycle.query.get(cycle.id)
       task1 = CycleTaskGroupObjectTask.query.get(
@@ -191,11 +191,11 @@ class TestCycleTaskStatusChange(TestCase):
 
       self.task_change_status(task1, "Finished")
 
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-02"):
-      notifications.send_todays_digest_notifications()
+      common.send_todays_digest_notifications()
 
       cycle = Cycle.query.get(cycle.id)
       task1 = CycleTaskGroupObjectTask.query.get(
@@ -204,7 +204,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.task_change_status(task1, "Declined")
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
 
       self.assertIn(user.email, notif_data)
       self.assertIn("task_declined", notif_data[user.email])
@@ -224,7 +224,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.wf_generator.activate_workflow(workflow)
 
     with freeze_time("2015-05-02"):
-      notifications.send_todays_digest_notifications()
+      common.send_todays_digest_notifications()
 
       cycle = Cycle.query.get(cycle.id)
       task1 = CycleTaskGroupObjectTask.query.get(
@@ -232,7 +232,7 @@ class TestCycleTaskStatusChange(TestCase):
 
       self.task_change_status(task1, "Finished")
 
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-03"):
@@ -243,7 +243,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.task_change_status(task1)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
       self.assertIn("all_tasks_completed", notif_data["user@example.com"])
 
@@ -260,7 +260,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.wf_generator.activate_workflow(workflow)
 
     with freeze_time("2015-05-03"):
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       cycle = Cycle.query.get(cycle.id)
       user = Person.query.get(self.user.id)
       self.assertIn(user.email, notif_data)
@@ -268,7 +268,7 @@ class TestCycleTaskStatusChange(TestCase):
       cycle = Cycle.query.get(cycle.id)
       self.assertFalse(cycle.is_current)
 
-      _, notif_data = notifications.get_todays_notifications()
+      _, notif_data = common.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
   def create_test_cases(self):
@@ -307,12 +307,6 @@ class TestCycleTaskStatusChange(TestCase):
             "contact": person_dict(self.user.id),
             "task_group_tasks": [{
                 "title": "task 1",
-                "description": "two taks in wf with different objects",
-                "contact": person_dict(self.user.id),
-                "start_date": date(2015, 5, 1),  # friday
-                "end_date": date(2015, 5, 5),
-            }, {
-                "title": "task 2",
                 "description": "two taks in wf with different objects",
                 "contact": person_dict(self.user.id),
                 "start_date": date(2015, 5, 1),  # friday
